Cleanup

+++ contracts/MarketSourceFactory.sol
@@ -14,7 +14,7 @@ contract MarketSourceFactory {
      * @param reportExpirationTimeSec The number of seconds after which the market data is deemed expired.
      * @return The address of the created MarketSource contract.
      */
-    function createSource(string name, uint256 reportExpirationTimeSec) public returns (MarketSource) {
+    function createSource(string name, uint64 reportExpirationTimeSec) public returns (MarketSource) {
         MarketSource source = new MarketSource(name, reportExpirationTimeSec);
         emit LogSourceCreated(msg.sender, source);
         source.transferOwnership(msg.sender);
